Rename components/DataViewTable.vue to components/index/_shared/DataViewTable.vue
